# swarm/plan/api_contracts.yaml
openapi: 3.0.3
info:
  title: Flow Studio SpecManager API
  description: |
    API contracts for the SpecManager service that powers the Flow Studio TypeScript frontend.

    This API provides:
    - **Spec Management**: CRUD operations for FlowGraphs and StepTemplates
    - **Validation**: Dry-run validation of spec changes
    - **Compilation**: Preview PromptPlan compilation from specs
    - **Run State**: Execution state management with GPS-style detour routing
    - **Event Streaming**: Server-Sent Events for real-time run updates

    ## Concurrency Model

    All mutable resources use **optimistic concurrency control** via ETags:
    - GET responses include `ETag` header with resource version hash
    - PATCH/PUT requests must include `If-Match` header with expected ETag
    - 409 Conflict returned if resource has been modified

    ## Error Handling

    All errors return `SpecError` with:
    - `error`: Machine-readable error code (e.g., `FLOW_NOT_FOUND`, `VALIDATION_FAILED`)
    - `message`: Human-readable description
    - `details`: Structured error context (optional)

  version: 2.0.0
  contact:
    name: Flow Studio Team
  license:
    name: MIT
servers:
  - url: /api
    description: Local development server
...
paths:
...
  /runs:
    post:
      operationId: createRun
      summary: Create a new run with initial state
      tags:
        - Runs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '201':
          description: Run created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunState'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecError'

    get:
      operationId: listRuns
      summary: List runs with optional filtering
      tags:
        - Runs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, succeeded, failed, canceled, paused, interrupted]
        - name: flow_key
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: since
          in: query
          description: Filter runs created after this timestamp
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of runs
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items:
                      $ref: '#/components/schemas/RunStateSummary'
                  total:
                    type: integer
                  has_more:
                    type: boolean
  /runs/{run_id}:
    parameters:
      - $ref: '#/components/parameters/RunId'

    get:
      operationId: getRunState
      summary: Get current run state
      tags:
        - Runs
      responses:
        '200':
          description: Current run state
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunState'
        '404':
          description: Run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecError'

    patch:
      operationId: updateRunState
      summary: Update run state (partial update)
      tags:
        - Runs
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunStateUpdate'
      responses:
        '200':
          description: Updated run state
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunState'
        '400':
          description: Invalid update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecError'
        '409':
          description: Conflict (state has changed since read)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecError'

  # ============================================================
  # RUN EVENT STREAMING
  # ============================================================
  /runs/{run_id}/events:
    parameters:
      - $ref: '#/components/parameters/RunId'

    get:
      operationId: streamRunEvents
      summary: Stream run events (SSE)
      description: |
        Opens a Server-Sent Events (SSE) stream for real-time run updates.

        **Event Types**:
        - `state_change`: Run status or step changed
        - `step_started`: Step execution began
        - `step_completed`: Step execution finished
        - `handoff_received`: Handoff envelope received
        - `routing_decision`: Routing decision made
        - `detour_started`: Detour sequence began
        - `detour_completed`: Detour sequence finished
        - `error`: Error occurred
        - `metrics_update`: Metrics updated

        **Reconnection**: Include `Last-Event-ID` header to resume from a specific event.
      tags:
        - Events
      parameters:
        - name: Last-Event-ID
          in: header
          description: Resume from this event ID
          schema:
            type: string
        - name: events
          in: query
          description: Filter event types (comma-separated)
          schema:
            type: string
          example: "state_change,step_completed"
      responses:
        '200':
          description: SSE event stream
          headers:
            Content-Type:
              schema:
                type: string
                enum: ["text/event-stream"]
            Cache-Control:
              schema:
                type: string
                enum: ["no-cache"]
            Connection:
              schema:
                type: string
                enum: ["keep-alive"]
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                event: state_change
                id: evt-001
                data: {"run_id":"run-20251228-100000-abc123","status":"running","current_step_id":"author_tests"}

                event: step_completed
                id: evt-002
                data: {"step_id":"author_tests","status":"VERIFIED","duration_ms":45000}

        '404':
          description: Run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecError'

  # ============================================================
  # RUN CONTROL ENDPOINTS
  # ============================================================
...
components:
...
  parameters:
...
    RunId:
      name: run_id
      in: path
      required: true
      description: Unique run identifier
      schema:
        type: string
        pattern: '^run-[0-9]{8}-[0-9]{6}-[a-z0-9]{6,8}$'
      example: "run-20251228-100000-abc123"
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: ETag from the most recent GET (for optimistic concurrency)
      schema:
        type: string
      example: '"abc123def456"'

  # ----------------------------------------------------------
  # HEADERS
  # ----------------------------------------------------------
  headers:
    ETag:
      description: Resource version hash for optimistic concurrency
      schema:
        type: string
      example: '"abc123def456"'

  # ----------------------------------------------------------
  # SCHEMAS
  # ----------------------------------------------------------
  schemas:
...
    CreateRunRequest:
      type: object
      required:
        - flow_key
      properties:
        flow_key:
          type: string
          description: Flow to execute
        capture_spec_versions:
          type: boolean
          default: true
          description: Whether to capture spec hashes at run start
        annotations:
          type: object
          additionalProperties: true
          description: User-defined annotations
    RunState:
      type: object
      description: |
        Full run state conforming to run_state.schema.json.
        Durable program counter for stepwise flow execution with GPS-style detour routing support.
      required:
        - run_id
        - flow_key
        - status
        - timestamp
      properties:
        run_id:
          type: string
        flow_key:
          type: string
        current_step_id:
          type: string
          nullable: true
        step_index:
          type: integer
        status:
          type: string
          enum: [pending, running, succeeded, failed, canceled, paused, interrupted]
        timestamp:
          type: string
          format: date-time
        loop_state:
          type: object
          additionalProperties:
            type: integer
        handoff_envelopes:
          type: object
        current_flow_index:
          type: integer
        flow_transition_history:
          type: array
          items:
            type: object
        interruption_stack:
          type: array
          items:
            $ref: '#/components/schemas/InterruptionFrame'
        resume_stack:
          type: array
          items:
            $ref: '#/components/schemas/ResumeFrame'
        spec_versions:
          $ref: '#/components/schemas/SpecVersions'
        injected_nodes:
          type: array
          items:
            $ref: '#/components/schemas/InjectedNode'
        routing_cursor:
          $ref: '#/components/schemas/RoutingCursor'
        metrics:
          $ref: '#/components/schemas/RunMetrics'
        annotations:
          type: object
          additionalProperties: true
    RunStateSummary:
      type: object
      description: Abbreviated run state for list views
      properties:
        run_id:
          type: string
        flow_key:
          type: string
        current_step_id:
          type: string
          nullable: true
        status:
          type: string
        timestamp:
          type: string
          format: date-time
        steps_completed:
          type: integer
        has_active_detour:
          type: boolean
    RunStateUpdate:
      type: object
      description: Partial update to run state
      properties:
        status:
          type: string
          enum: [pending, running, succeeded, failed, canceled, paused, interrupted]
        current_step_id:
          type: string
        step_index:
          type: integer
        annotations:
          type: object
          additionalProperties: true
    InterruptionFrame:
      type: object
      required:
        - id
        - type
        - timestamp
        - trigger
      properties:
        id:
          type: string
          pattern: '^int-[a-z0-9]{8}$'
        type:
          type: string
          enum: [human_injection, policy_enforcement, error_recovery, priority_override, branch_detour, external_event]
        timestamp:
          type: string
          format: date-time
        trigger:
          type: object
          required:
            - source
          properties:
            source:
              type: string
              enum: [operator, policy_engine, step_output, external_api, timeout, signal]
            source_id:
              type: string
            payload:
              type: object
        detour_steps:
          type: array
          items:
            type: string
        current_detour_index:
          type: integer
          default: 0
        priority:
          type: integer
          minimum: 0
          maximum: 100
          default: 50
        can_abort:
          type: boolean
          default: true
        completion_condition:
          type: object
    ResumeFrame:
      type: object
      required:
        - flow_key
        - step_id
        - step_index
        - interruption_id
      properties:
        flow_key:
          type: string
        step_id:
          type: string
        step_index:
          type: integer
        interruption_id:
          type: string
        loop_state_snapshot:
          type: object
          additionalProperties:
            type: integer
        resume_mode:
          type: string
          enum: [continue, retry, skip, conditional]
          default: continue
    SpecVersions:
      type: object
      properties:
        flow_hash:
          type: string
          pattern: '^[a-f0-9]{64}$'
        station_hashes:
          type: object
          additionalProperties:
            type: string
        template_hashes:
          type: object
          additionalProperties:
            type: string
        prompt_hashes:
          type: object
          additionalProperties:
            type: string
        captured_at:
          type: string
          format: date-time
        drift_detected:
          type: boolean
          default: false
    InjectedNode:
      type: object
      required:
        - id
        - type
        - injected_at
        - source
      properties:
        id:
          type: string
          pattern: '^inj-[a-z0-9]{8}$'
        type:
          type: string
          enum: [step, checkpoint, gate, fork, join, conditional]
        injected_at:
          type: string
          format: date-time
        source:
          type: string
          enum: [operator, policy, api, self_heal, auto_scale]
        position:
          type: object
          properties:
            relative_to:
              type: string
            placement:
              type: string
              enum: [before, after, replace, parallel]
        executed:
          type: boolean
          default: false
        reason:
          type: string
    RoutingCursor:
      type: object
      properties:
        decision_history:
          type: array
          items:
            type: object
        pending_branches:
          type: array
          items:
            type: object
        route_overrides:
          type: object
          additionalProperties:
            type: string
        blacklisted_steps:
          type: array
          items:
            type: string
        preferred_route:
          type: array
          items:
            type: string
    RunMetrics:
      type: object
      properties:
        total_duration_ms:
          type: integer
          minimum: 0
        steps_completed:
          type: integer
          minimum: 0
        steps_failed:
          type: integer
          minimum: 0
        steps_skipped:
          type: integer
          minimum: 0
        total_iterations:
          type: integer
          minimum: 0
        detours_taken:
          type: integer
          minimum: 0
        injections_processed:
          type: integer
          minimum: 0
        token_usage:
          type: object
          properties:
            input_tokens:
              type: integer
            output_tokens:
              type: integer
            total_tokens:
              type: integer
        cost_estimate_usd:
          type: number
          minimum: 0
...
    SpecError:
      type: object
      description: Standard error response for spec operations
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Machine-readable error code
          enum:
            - FLOW_NOT_FOUND
            - TEMPLATE_NOT_FOUND
            - RUN_NOT_FOUND
            - VALIDATION_FAILED
            - COMPILATION_FAILED
            - ETAG_MISMATCH
            - PRECONDITION_FAILED
            - INVALID_PATCH
            - INVALID_REQUEST
            - CONFLICT
            - INTERNAL_ERROR
        message:
          type: string
          description: Human-readable error description
        details:
          type: object
          additionalProperties: true
          description: Additional error context
...